Rename log4j module to log4j2 for consistency with the actual Log4j2 library
The module name, package, class names, and property prefix all used
"log4j" but the library targets Log4j2. Rename everything to "log4j2"
to avoid confusion.
