--- otel-logs-autoconfigure-log4j/src/main/resources/META-INF/additional-spring-configuration-metadata.json
@@ -1,49 +1,0 @@
-{
-  "groups": [
-    {
-      "name": "management.opentelemetry.instrumentation.log4j-appender"
-    }
-  ],
-  "properties": [
-    {
-      "name": "management.opentelemetry.instrumentation.log4j-appender.capture-code-attributes",
-      "type": "java.lang.Boolean",
-      "defaultValue": false
-    },
-    {
-      "name": "management.opentelemetry.instrumentation.log4j-appender.capture-context-data-attributes",
-      "type": "java.lang.String"
-    },
-    {
-      "name": "management.opentelemetry.instrumentation.log4j-appender.capture-event-name",
-      "type": "java.lang.Boolean",
-      "defaultValue": false
-    },
-    {
-      "name": "management.opentelemetry.instrumentation.log4j-appender.capture-experimental-attributes",
-      "type": "java.lang.Boolean",
-      "defaultValue": false
-    },
-    {
-      "name": "management.opentelemetry.instrumentation.log4j-appender.capture-map-message-attributes",
-      "type": "java.lang.Boolean",
-      "defaultValue": false
-    },
-    {
-      "name": "management.opentelemetry.instrumentation.log4j-appender.capture-marker-attribute",
-      "type": "java.lang.Boolean",
-      "defaultValue": false
-    },
-    {
-      "name": "management.opentelemetry.instrumentation.log4j-appender.enabled",
-      "type": "java.lang.Boolean",
-      "defaultValue": true
-    },
-    {
-      "name": "management.opentelemetry.instrumentation.log4j-appender.num-logs-captured-before-otel-install",
-      "type": "java.lang.Integer",
-      "defaultValue": 1000
-    }
-  ],
-  "hints": []
-}
