+++ otel-logs-autoconfigure-log4j2/src/test/resources/log4j2-test.xml
